Split MemoryTracker header
Allocator.h is included into almost everything (via PODArray.h), so keep
only small part of MemoryTracker.h for it.
